탭바 적용을 완료했다.
하지만 두가지 문제가 있다.

+++ CustomCellTable/CartCell.m
@@ -40,7 +40,7 @@
     
     // 셀 반영
     self.name.text = item.product.name;
-    self.quantity.text = [NSString stringWithFormat:@"%d개", item.quantity];
+    self.quantity.text = [NSString stringWithFormat:@"%d개", (int)item.quantity];
 }
 
 @end
